# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mattermost RTK Plugin API
  license:
    name: MIT
    url: https://github.com/kondo97/mattermost-plugin-rtk/blob/master/LICENSE
  description: |
    HTTP API for the plugin that integrates Cloudflare RealtimeKit (RTK) into Mattermost channel video and audio calls.

    ## Authentication

    ### Mattermost Authentication
    The Mattermost internal HTTP client automatically attaches the `Mattermost-User-ID` header.
    External clients must provide a valid session cookie or Personal Access Token.

    ### RTK Webhook Authentication
    Event notifications from RTK include an HMAC-SHA256 signature in the `dyte-signature` header.
    The signature is verified using the webhook secret held by the plugin.
  version: "1.0.0"
  contact:
    url: https://github.com/kondo97/mattermost-plugin-rtk/issues

servers:
  - url: /plugins/com.kondo97.mattermost-plugin-rtk
    description: Mattermost plugin base path (production)

tags:
  - name: Calls
    description: Create, join, leave, and end calls
  - name: Channels
    description: Channel call state (Calls-plugin compatible)
  - name: Config
    description: Retrieve plugin configuration status
  - name: Mobile
    description: Mobile-specific operations
  - name: Webhook
    description: Event notifications from RTK

paths:
  /api/v1/calls:
    post:
      operationId: createCall
      tags: [Calls]
      summary: Start a call
      description: |
        Creates a new call in the specified channel.
        Returns 409 if there is already an active call in the channel.
        The creator is automatically added as a participant.
      security:
        - MattermostAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCallRequest'
            example:
              channel_id: "abc123def456ghi789"
      responses:
        "201":
          description: Call created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallWithToken'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          $ref: '#/components/responses/Conflict'
        "500":
          $ref: '#/components/responses/InternalServerError'
        "503":
          $ref: '#/components/responses/ServiceUnavailable'

  /api/v1/calls/{id}:
    get:
      operationId: getCall
      tags: [Calls]
      summary: Get call information
      description: Returns the call session information for the specified ID.
      security:
        - MattermostAuth: []
      parameters:
        - $ref: '#/components/parameters/CallID'
      responses:
        "200":
          description: Call session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallSession'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'

    delete:
      operationId: endCall
      tags: [Calls]
      summary: End a call
      description: |
        Ends the specified call.
        Only the call creator can perform this action. Other users receive 403.
      security:
        - MattermostAuth: []
      parameters:
        - $ref: '#/components/parameters/CallID'
      responses:
        "200":
          description: Call ended
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'

  /api/v1/calls/{id}/token:
    post:
      operationId: joinCall
      tags: [Calls]
      summary: Join a call (get token)
      description: |
        Issues an RTK authentication token to join an existing call.
        The user is added to the participant list (duplicates are ignored).
      security:
        - MattermostAuth: []
      parameters:
        - $ref: '#/components/parameters/CallID'
      responses:
        "200":
          description: Join token and call session information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallWithToken'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
        "503":
          $ref: '#/components/responses/ServiceUnavailable'

  /api/v1/calls/{id}/leave:
    post:
      operationId: leaveCall
      tags: [Calls]
      summary: Leave a call
      description: |
        Leaves the call. If the last participant leaves, the call ends automatically.
        Idempotent -- no error is returned if already left or the call has already ended.
      security:
        - MattermostAuth: []
      parameters:
        - $ref: '#/components/parameters/CallID'
      responses:
        "200":
          description: Left successfully
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalServerError'

  /api/v1/calls/{id}/dismiss:
    post:
      operationId: dismissNotification
      tags: [Mobile]
      summary: Dismiss a call notification
      description: |
        Mobile-specific endpoint.
        Sends a `notification_dismissed` WebSocket event only to the requesting user.
        The call state is not changed (idempotent).
      security:
        - MattermostAuth: []
      parameters:
        - $ref: '#/components/parameters/CallID'
      responses:
        "200":
          description: Notification dismissed
        "401":
          $ref: '#/components/responses/Unauthorized'

  /api/v1/config/status:
    get:
      operationId: getConfigStatus
      tags: [Config]
      summary: Get plugin enabled status
      description: Returns whether RTK credentials are configured. Feature on/off is managed by the Cloudflare RTK Preset.
      security:
        - MattermostAuth: []
      responses:
        "200":
          description: Configuration status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigStatus'
              example:
                enabled: true
        "401":
          $ref: '#/components/responses/Unauthorized'

  /api/v1/config/admin-status:
    get:
      operationId: getAdminConfigStatus
      tags: [Config]
      summary: Get detailed admin configuration status
      description: |
        Accessible only to users with system administrator privileges.
        Returns detailed information such as how credentials are configured (environment variables / system console).
        **Note**: The response does not include secret values (API key itself).
      security:
        - MattermostAuth: []
      responses:
        "200":
          description: Admin configuration status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminConfigStatus'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'

  /api/v1/channels:
    get:
      operationId: getAllChannels
      tags: [Channels]
      summary: Get all channel call states
      description: |
        Returns one entry per channel for which the requesting user has channel membership and either:
        - A row exists in `rtk_calls_channels` (explicitly registered), or
        - An active call exists in the channel.

        The response shape mirrors the Calls plugin's `GET /channels` endpoint so existing Calls clients
        can be repointed at the RTK plugin without changes.
      security:
        - MattermostAuth: []
      responses:
        "200":
          description: Array of channel call states
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChannelCallState'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalServerError'

  /api/v1/channels/{channelID}:
    get:
      operationId: getChannel
      tags: [Channels]
      summary: Get a single channel call state
      description: |
        Returns the calls-enabled flag for a single channel.
        If no row exists in `rtk_calls_channels` for the channel, it is treated as enabled by default.
      security:
        - MattermostAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelID'
      responses:
        "200":
          description: Channel call state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelStateResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalServerError'

    put:
      operationId: updateChannel
      tags: [Channels]
      summary: Update channel calls-enabled flag
      description: |
        Sets the calls-enabled flag for the channel. Only channel admins, team
        admins, and system admins may call this endpoint; other users receive 403.
      security:
        - MattermostAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChannelRequest'
            example:
              enabled: true
      responses:
        "200":
          description: Channel updated. Returns the channel's new calls-enabled state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelStateResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/InternalServerError'

  /api/v1/webhook/rtk:
    post:
      operationId: handleRTKWebhook
      tags: [Webhook]
      summary: Receive RTK webhook events
      description: |
        Endpoint that receives event notifications from Cloudflare RTK.

        Supported events:
        - `meeting.participantJoined` -- a participant entered the meeting (used to capture `session_id`)
        - `meeting.participantLeft` -- a participant left the meeting
        - `meeting.ended` -- the meeting ended

        Unknown event types are ignored.

        > **Note**: HMAC-SHA256 signature verification (`dyte-signature` header) is not yet
        > implemented. The endpoint currently accepts all POST requests without verifying
        > the signature.
      security:
        - RTKSignature: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RTKWebhookEvent'
            examples:
              participantJoined:
                summary: Participant joined event
                value:
                  event: "meeting.participantJoined"
                  meeting:
                    id: "meeting-uuid"
                    sessionId: "session-uuid"
                  participant:
                    customParticipantId: "user-mattermost-id"
              meetingEnded:
                summary: Meeting ended event
                value:
                  event: "meeting.ended"
                  meeting:
                    id: "meeting-uuid"
                  participant:
                    customParticipantId: ""
      responses:
        "200":
          description: Event accepted
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    MattermostAuth:
      type: apiKey
      in: header
      name: Mattermost-User-ID
      description: |
        Mattermost internal authentication. Browser clients attach this automatically via session cookie.
        The value is the authenticated user's Mattermost User ID.
    RTKSignature:
      type: apiKey
      in: header
      name: dyte-signature
      description: |
        HMAC-SHA256 signature header sent by RTK.
        Format: `hex(HMAC-SHA256(webhook_secret, raw_body))`.
        **Note**: Signature verification is not yet implemented in the plugin.

  parameters:
    CallID:
      name: id
      in: path
      required: true
      description: UUID of the call
      schema:
        type: string
        format: uuid
        example: "550e8400-e29b-41d4-a716-446655440000"
    ChannelID:
      name: channelID
      in: path
      required: true
      description: ID of the Mattermost channel
      schema:
        type: string
        example: "abc123def456ghi789"

  schemas:
    CreateCallRequest:
      type: object
      required: [channel_id]
      properties:
        channel_id:
          type: string
          description: ID of the Mattermost channel to create the call in
          example: "abc123def456ghi789"

    CallSession:
      type: object
      description: State of a call session
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the call
        channel_id:
          type: string
          description: ID of the Mattermost channel the call belongs to
        creator_id:
          type: string
          description: Mattermost User ID of the user who created the call
        meeting_id:
          type: string
          description: Cloudflare RTK meeting ID
        participants:
          type: array
          items:
            type: string
          description: List of Mattermost User IDs of current participants
        create_at:
          type: integer
          format: int64
          description: Call start time (Unix milliseconds)
        update_at:
          type: integer
          format: int64
          description: Last update time (Unix milliseconds)
        end_at:
          type: integer
          format: int64
          description: Call end time (Unix milliseconds). 0 indicates the call is active
        post_id:
          type: string
          description: ID of the call notification post published to the channel
        rtk_channel_meeting_id:
          type: string
          description: ID of the rtk_channel_meetings row this call was created against
        session_id:
          type: string
          description: Cloudflare RTK session UUID. Populated via webhook when the first participant connects; empty string if the webhook has not yet been received.
      example:
        id: "550e8400-e29b-41d4-a716-446655440000"
        channel_id: "abc123def456ghi789"
        creator_id: "user001"
        meeting_id: "rtk-meeting-id"
        participants: ["user001", "user002"]
        create_at: 1700000000000
        update_at: 1700000001000
        end_at: 0
        post_id: "post001"
        rtk_channel_meeting_id: "mtg001"
        session_id: "05e57591-d89e-4c2a-b3f1-123456789abc"

    CallWithToken:
      type: object
      description: Call session and RTK authentication token
      properties:
        call:
          $ref: '#/components/schemas/CallSession'
        token:
          type: string
          description: JWT token used to connect to the RTK SDK

    ConfigStatus:
      type: object
      properties:
        enabled:
          type: boolean
          description: Whether RTK credentials are configured and the plugin is available

    AdminConfigStatus:
      type: object
      properties:
        enabled:
          type: boolean
          description: Whether RTK credentials are configured and the plugin is available
        account_id_via_env:
          type: boolean
          description: Whether the Cloudflare Account ID is configured via environment variable
        api_token_via_env:
          type: boolean
          description: Whether the Cloudflare API Token is configured via environment variable
        app_id_via_env:
          type: boolean
          description: Whether the RTK App ID is pinned via the RTK_APP_ID environment variable
        cloudflare_account_id:
          type: string
          description: Cloudflare Account ID configured in the system console

    ChannelStateResponse:
      type: object
      description: Calls-enabled state for a single channel.
      required: [channel_id, enabled]
      properties:
        channel_id:
          type: string
          description: Mattermost channel ID
        enabled:
          type: boolean
          description: |
            Whether calls are enabled for the channel.
            Defaults to `true` when no row exists in `rtk_calls_channels`.
      example:
        channel_id: "abc123def456ghi789"
        enabled: true

    UpdateChannelRequest:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
          description: New calls-enabled flag for the channel

    ChannelCallState:
      type: object
      description: |
        One element of the GET /api/v1/channels response array.
        Mirrors the Calls plugin's channel call state shape for client compatibility.
      properties:
        channel_id:
          type: string
          description: Mattermost channel ID
        enabled:
          type: boolean
          nullable: true
          description: Whether calls are explicitly enabled for the channel. Omitted when the entry was synthesised from an active call with no rtk_calls_channels row.
        call:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/CallStateClient'
          description: Active call state, or null if no call is in progress

    CallStateClient:
      type: object
      description: Per-call payload returned inside ChannelCallState. Field names mirror the Calls plugin wire format.
      properties:
        id:
          type: string
          format: uuid
          description: Unique call identifier
        start_at:
          type: integer
          format: int64
          description: Call start time (Unix milliseconds)
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/UserSessionClient'
          description: Per-participant session state
        thread_id:
          type: string
          description: Thread ID (empty — RTK does not use threaded posts)
        post_id:
          type: string
          description: ID of the call notification post
        screen_sharing_session_id:
          type: string
          description: Screen-sharing session ID (empty — not yet supported)
        owner_id:
          type: string
          description: Mattermost User ID of the call creator
        host_id:
          type: string
          description: Mattermost User ID of the current host (same as owner_id)
        recording:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/JobStateClient'
          description: Recording job state (always null — not supported)
        transcription:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/JobStateClient'
          description: Transcription job state (always null — not supported)
        live_captions:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/JobStateClient'
          description: Live captions job state (always null — not supported)
        dismissed_notification:
          type: object
          nullable: true
          additionalProperties:
            type: boolean
          description: Map of user IDs who have dismissed the call notification

    UserSessionClient:
      type: object
      description: Per-participant session state. RTK does not track per-session mute/hand/video state; boolean fields are always false and session_id equals user_id.
      properties:
        session_id:
          type: string
          description: Session identifier (set to user_id as a placeholder)
        user_id:
          type: string
          description: Mattermost User ID
        unmuted:
          type: boolean
          description: Whether the participant is unmuted (always false)
        raised_hand:
          type: integer
          format: int64
          description: Raised-hand timestamp (always 0)
        video:
          type: boolean
          description: Whether the participant has video enabled (always false)

    JobStateClient:
      type: object
      description: State of a background job (recording, transcription, or live captions). RTK does not run these jobs; this type exists solely for Calls-plugin wire compatibility.
      properties:
        type:
          type: string
        init_at:
          type: integer
          format: int64
        start_at:
          type: integer
          format: int64
        end_at:
          type: integer
          format: int64
        err:
          type: string

    RTKWebhookEvent:
      type: object
      description: Payload of a webhook event sent from RTK
      required: [event, meeting, participant]
      properties:
        event:
          type: string
          description: Event type
          enum:
            - meeting.participantJoined
            - meeting.participantLeft
            - meeting.ended
        meeting:
          type: object
          properties:
            id:
              type: string
              description: RTK meeting ID
            sessionId:
              type: string
              description: RTK session UUID (one per active call instance). Present in all event types.
        participant:
          type: object
          properties:
            customParticipantId:
              type: string
              description: Mattermost User ID of the participant

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      example:
        error: "call not found or already ended"

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "channel_id is required"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Not authorized"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "only the call creator can perform this action"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "call not found or already ended"
    Conflict:
      description: There is already an active call in the channel
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "call already active in channel"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "internal error"
    ServiceUnavailable:
      description: RTK credentials are not configured
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "cloudflare RTK credentials are not configured"